[TESTS] Separates events tests cases from user test cases
